fix 'sample' bug
This commit fixes the sampling bug for most cases. It will look for the most abundant class (var type) in the df, and then select from that, to avoid randomly selecting a single or double variable type. In a small df, this could still be a problem, but in a sufficiently large dataframe, this should fix the small likliehood that this happens.